admin-tools/requirements.txt
PyJWT>=2.0.0
requests>=2.25.0
cryptography>=3.4.7
